+++ project-mobile-app/App.js
@@ -1,23 +1,15 @@
-// Importing the Modules
-import * as React from "react";
-import { useState } from "react";
-import { StyleSheet, View, Image } from "react-native";
-import { CalendarList } from "react-native-calendars";
-// import Day from "react-native-calendars/src/calendar/day/basic";
-
+import React from "react";
 import { NavigationContainer } from "@react-navigation/native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import { createMaterialTopTabNavigator } from "@react-navigation/material-top-tabs";
 
 import HomeScreen from "./screen/HomeScreen";
-// import AgendaScreen from "./screen/AgendaScreen";
 
 const Stack = createNativeStackNavigator();
 
 function MyStack() {
   return (
     <Stack.Navigator>
-      <Stack.Screen name="Home" component={HomeScreen} />
+      <Stack.Screen name="Calendar" component={HomeScreen} />
     </Stack.Navigator>
   );
 }
