+++ project-mobile-app/screen/HomeScreen.Archive.js
@@ -1,0 +1,132 @@
+import React, { Component } from "react";
+import { Alert, StyleSheet, Text, View, TouchableOpacity } from "react-native";
+import { Agenda } from "react-native-calendars";
+import testIDs from "../testIDs";
+
+// export default: 
+// This syntax exports the HomeScreen class so it can be used in other parts of the application. When a file contains a default export, the importing file can choose any name to refer to the exported class or function.
+// class HomeScreen extends Component: This defines a class called AgendaScreen that extends from Component, meaning AgendaScreen is a React component. 
+// Component is the base class provided by React that provides methods and lifecycle hooks like render(), componentDidMount(), etc.
+
+// state: This is where the component’s local state is defined. React components use state to store information that can change over time, such as user input or data fetched from an API.
+// items: undefined: This is the initial state value for a property called items. In this case, it's being initialized as undefined, meaning that items doesn't yet have any value. 
+// This might later be populated with some data, such as a list of items, likely to be displayed in the AgendaScreen.
+
+export default class HomeScreen extends Component {
+  state = {
+    items: undefined,
+  };
+
+  // The render() function you’ve shared is rendering an Agenda component, likely from the react-native-calendars library
+  render() {
+    return (
+      <Agenda
+        pastScrollRange={1}
+        futureScrollRange={1}
+        testID={testIDs.agenda.CONTAINER}
+        items={this.state.items}
+        loadItemsForMonth={this.loadItems}
+        selected={"2025-01-16"}
+        renderItem={this.renderItem}
+        renderEmptyDate={this.renderEmptyDate}
+        rowHasChanged={this.rowHasChanged}
+        showClosingKnob={true}
+      />
+    );
+  }
+
+  loadItems = (day) => {
+    const items = this.state.items || {};
+
+    setTimeout(() => {
+      for (let i = -15; i < 85; i++) {
+        const time = day.timestamp + i * 24 * 60 * 60 * 1000;
+        const strTime = this.timeToString(time);
+
+        if (!items[strTime]) {
+          items[strTime] = [];
+
+          const numItems = Math.floor(Math.random() * 3 + 1);
+          for (let j = 0; j < numItems; j++) {
+            items[strTime].push({
+              name: "Item for " + strTime + " #" + j,
+              height: Math.max(50, Math.floor(Math.random() * 150)),
+              day: strTime,
+            });
+          }
+        }
+      }
+
+      const newItems = {};
+      Object.keys(items).forEach((key) => {
+        newItems[key] = items[key];
+      });
+      this.setState({
+        items: newItems,
+      });
+    }, 10);
+  };
+
+  renderDay = (day) => {
+    if (day) {
+      return <Text style={styles.customDay}>{day.getDay()}</Text>;
+    }
+    return <View style={styles.dayItem} />;
+  };
+
+  renderItem = (reservation, isFirst) => {
+    const fontSize = isFirst ? 16 : 14;
+    const color = isFirst ? "black" : "#43515c";
+
+    return (
+      <TouchableOpacity
+        testID={testIDs.agenda.ITEM}
+        style={[styles.item, { height: reservation.height }]}
+        onPress={() => Alert.alert(reservation.name)}
+      >
+        <Text style={{ fontSize, color }}>{reservation.name}</Text>
+      </TouchableOpacity>
+    );
+  };
+
+  renderEmptyDate = () => {
+    return (
+      <View style={styles.emptyDate}>
+        <Text>This is empty date!</Text>
+      </View>
+    );
+  };
+
+  rowHasChanged = (r1, r2) => {
+    return r1.name !== r2.name;
+  };
+
+  timeToString(time) {
+    const date = new Date(time);
+    return date.toISOString().split("T")[0];
+  }
+}
+
+const styles = StyleSheet.create({
+  item: {
+    backgroundColor: "white",
+    flex: 1,
+    borderRadius: 5,
+    padding: 10,
+    marginRight: 10,
+    marginTop: 17,
+  },
+  emptyDate: {
+    height: 15,
+    flex: 1,
+    paddingTop: 30,
+  },
+  customDay: {
+    margin: 10,
+    fontSize: 24,
+    color: "green",
+  },
+  dayItem: {
+    marginLeft: 34,
+  },
+});
